use proper schema title
