+++ test/schemaValidation.test.ts
@@ -1634,7 +1634,7 @@ obj:
       const content = `foo: bar`;
       const result = await parseSetup(content);
       expect(result).to.have.length(1);
-      expect(result[0].message).to.include("Schema 'undefined' is not valid");
+      expect(result[0].message).to.include("Schema 'default_schema_id.yaml' is not valid");
       expect(telemetry.messages).to.be.empty;
     });
